Refactor: move `ApiJsonObjectMigrationBuilder` to separate file
This is in preparation of a new needed `ApiJsonArrayMigrationBuilder`.
