+++ carp.common/src/commonMain/kotlin/dk/cachet/carp/common/infrastructure/versioning/ApiMigration.kt
@@ -38,8 +38,8 @@ abstract class ApiMigration( val minimumMinorVersion: Int, val targetMinorVersio
     abstract fun migrateResponse( request: JsonObject, response: ApiResponse, targetVersion: ApiVersion ): ApiResponse
     abstract fun migrateEvent( event: JsonObject ): JsonObject
 
-    protected fun JsonObject.migrate( migration: ApiMigrationBuilder.() -> Unit ): JsonObject =
-        ApiMigrationBuilder( this, minimumMinorVersion, targetMinorVersion )
+    protected fun JsonObject.migrate( migration: ApiJsonObjectMigrationBuilder.() -> Unit ): JsonObject =
+        ApiJsonObjectMigrationBuilder( this, minimumMinorVersion, targetMinorVersion )
             .apply( migration ).build()
 }
 
@@ -57,58 +57,4 @@ class ApiResponse( val response: JsonElement?, val ex: Exception? )
 }
 
 
-class ApiMigrationBuilder(
-    jsonObject: Map<String, JsonElement>,
-    val minimumMinorVersion: Int,
-    val targetMinorVersion: Int
-)
-{
-    val json: MutableMap<String, JsonElement> = jsonObject.toMutableMap()
 
-    init
-    {
-        // When API version is included, it always needs to be migrated.
-        val version: JsonElement? = json[ API_VERSION_FIELD ]
-        if ( version != null )
-        {
-            json[ API_VERSION_FIELD ] = version.replaceString( ".$minimumMinorVersion", ".$targetMinorVersion" )
-        }
-    }
-
-
-    /**
-     * Run a [migration] on this object in case it is a polymorphic type identified by [classDiscriminator].
-     */
-    fun ifType( classDiscriminator: String, migration: ApiMigrationBuilder.() -> Unit )
-    {
-        if ( json.getType() == classDiscriminator ) apply( migration )
-    }
-
-    /**
-     * Retrieve the object identified by [fieldName] in this object, and run the specified [migration].
-     */
-    fun updateObject( fieldName: String, migration: ApiMigrationBuilder.() -> Unit )
-    {
-        val o = requireNotNull( json[ fieldName ] as? JsonObject )
-        val newJson: JsonObject = ApiMigrationBuilder( o, minimumMinorVersion, targetMinorVersion )
-            .apply( migration ).build()
-        json[ fieldName ] = newJson
-    }
-
-    /**
-     * Copy the value of the field with [fromFieldName] to the field with [toFieldName].
-     */
-    fun copyField( fromFieldName: String, toFieldName: String )
-    {
-        val from = requireNotNull( json[ fromFieldName ] )
-        json[ toFieldName ] = from
-    }
-
-    private fun JsonElement.replaceString( oldValue: String, newValue: String ): JsonPrimitive
-    {
-        require( this is JsonPrimitive && this.isString )
-        return JsonPrimitive( content.replace( oldValue, newValue ) )
-    }
-
-    fun build(): JsonObject = JsonObject( json.toMap() )
-}
